core: Move sources from bindings to core part
