+++ include/pinocchio/bindings/python/parsers/python.hpp
@@ -5,39 +5,12 @@
 #ifndef __pinocchio_python_parser_python_hpp__
 #define __pinocchio_python_parser_python_hpp__
 
-#include "pinocchio/multibody/model.hpp"
+#include "pinocchio/macros.hpp"
 
-#include <boost/python.hpp>
+// clang-format off
+PINOCCHIO_PRAGMA_DEPRECATED_HEADER(pinocchio/bindings/python/parsers/python.hpp, pinocchio/parsers/python.hpp)
+// clang-format on
 
-#if defined _WIN32
-  #ifdef pinocchio_pywrap_EXPORTS
-    #define PINOCCHIO_PYWRAP_DLLAPI __declspec(dllexport)
-  #else
-    #define PINOCCHIO_PYWRAP_DLLAPI __declspec(dllimport)
-  #endif // pinocchio_pywrap_EXPORTS
-#else
-  #define PINOCCHIO_PYWRAP_DLLAPI
-#endif // _WIN32
-
-namespace pinocchio
-{
-  namespace python
-  {
-    /// \brief Load a model from a Python script.
-    ///
-    /// This function raises a Python error in case of incistency in the Python code.
-    ///
-    /// \input filename The full path to the model file.
-    /// \input var_name Name of the Python variable which contains the model in the script.
-    ///
-    /// \returns The model constructed by the Python script.
-    ///
-    // TODO: look inside the context of Python and find an occurence of object Model
-    PINOCCHIO_PYWRAP_DLLAPI
-    Model buildModel(const std::string & filename, const std::string & var_name = "model");
-
-  } // namespace python
-
-} // namespace pinocchio
+#include "pinocchio/parsers/python.hpp"
 
 #endif // ifndef __pinocchio_python_parser_python_hpp__
